backend/requirements.txt
kafka-python
faker
requests
Flask
Flask-SocketIO
Flask-Cors
SQLAlchemy
PyJWT
certifi
python-dotenv
openai